+++ ts/packages/providers/llamaindex/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @composio/llamaindex
 
+## 0.2.4
+
+### Patch Changes
+
+- Updated dependencies [97c4138]
+  - @composio/core@0.2.4
+
+## 0.2.3
+
+### Patch Changes
+
+- Updated dependencies [f88ab99]
+  - @composio/core@0.2.3
+
+## 0.2.2
+
+### Patch Changes
+
+- cfc2c50: Update zod version to 4
+- Updated dependencies [cfc2c50]
+  - @composio/core@0.2.2
+
 ## 0.2.1
 
 ### Patch Changes